requirements.txt
mysql-connector-python==8.0.27
influxdb==5.3.1